requirements.txt
tensorflow==2.16.1
ml-dtypes==0.3.1
jax==0.4.20
jaxlib==0.4.20
numpy>=1.22.0
opencv-python>=4.5.5.64
imutils>=0.5.4
scikit-learn>=1.0.2
pillow>=9.0.0